refactor: extract token and git collectors from usage-manager (#195)
